flask_app/requirements.txt
Flask==0.11.1
djaodjin-deployutils==0.2.9
